requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.34.2
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.15.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.7
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.9.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2023.10.0
gdown==5.1.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-aiplatform==1.66.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imageio==2.34.1
ipdb==0.13.13
ipykernel==6.29.3
ipython==8.22.2
jax==0.4.31
jaxlib==0.4.31
jaxtyping==0.2.34
jedi==0.19.1
Jinja2==3.1.3
jiter==0.5.0
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.44.1
packaging==24.0
pandarallel==1.6.5
pandas==2.2.1
parso==0.8.3
persist-to-disk==0.0.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
portalocker==2.10.1
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==5.28.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
responses==0.18.0
rouge-score==0.1.2
rsa==4.9
sacrebleu==2.4.3
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.6
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.4.0
tifffile==2024.4.18
tiktoken==0.7.0
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
triton==2.2.0
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
